`render_tag_element` accepts `&dyn TagRenderable`
This is because we are about to do the same thing we did for attributes
where our returning of a dyn trait means that the function as it calls
itself recursively will need to accept this as well. This means we need
to have the name method have a receiver to make all of the v table
lookups happy.