Introduce GenericElement with first one <h1>
This involves some public/private fiddling as we need this
GenericElement to be public, but it requires all elements to also be the
TagRenderable which returns the Attributes, so they all move at once.
Again the <h1> addition is mean to be simple enough that the rest of the
complexity should feel isolated.