Convert String-containing styles to &'static str
This was causing lifetime issues with the `(...).to_string()` that I
was forced to use to convert the str to a String. If we need to make
these parameterized over non-static strs, we can cross that bridge when
we get there. No one seems to be using these with `String` values yet,
so no changes need on any consumers of this.